requirements.txt
openai>=1.0.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-openai>=0.0.5
pypdf2>=3.0.0
pymupdf>=1.23.0
python-pptx>=0.6.21
docx2txt>=0.8
sentence-transformers>=2.2.0
tiktoken>=0.5.0
numpy>=1.22.4
pandas>=1.5.3
tqdm>=4.65.0
pillow>=9.0.0
pytesseract>=0.3.10
streamlit>=1.28.0
rank-bm25>=0.2.2
jieba>=0.42.1
